[TEST] Renamed integration specs using `*IntegrationSpec` as a suffix.
